fix numpy backend
