+++ funsor/interpreter.py
@@ -15,7 +15,7 @@ import numpy as np
 from funsor.domains import Domain
 from funsor.ops import Op, is_tensor
 from funsor.registry import KeyedRegistry
-#from funsor.tensor import is_nn_module
+from funsor.util import is_nn_module
 
 _ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 _DEBUG = int(os.environ.get("FUNSOR_DEBUG", 0))
@@ -153,7 +153,9 @@ _ground_types = (
     types.BuiltinFunctionType,
     Domain,
     Op,
-    np.ufunc
+    np.generic,
+    np.ndarray,
+    np.ufunc,
 )
 
 
@@ -218,7 +220,7 @@ for t in _ground_types:
 def is_atom(x):
     if isinstance(x, (tuple, frozenset)) and not isinstance(x, Domain):
         return len(x) == 0 or all(is_atom(c) for c in x)
-    return isinstance(x, _ground_types) or is_tensor(x) #or is_nn_module(x)
+    return isinstance(x, _ground_types) or is_tensor(x) or is_nn_module(x)
 
 
 def gensym(x=None):
